added functionality to gallery page
